RANGER-671 : Add support to retrieve permissions for the logged in user from UserSession rather going to database every time

+++ security-admin/src/main/java/org/apache/ranger/rest/XUserREST.java
@@ -761,6 +761,7 @@ public class XUserREST {
 	@Produces({ "application/xml", "application/json" })
 	@PreAuthorize("@rangerPreAuthSecurityHandler.isAPIAccessible(\"" + RangerAPIList.CREATE_X_MODULE_DEF_PERMISSION + "\")")
 	public VXModuleDef createXModuleDefPermission(VXModuleDef vXModuleDef) {
+		xUserMgr.checkAdminAccess();
 		return xUserMgr.createXModuleDefPermission(vXModuleDef);
 	}
 
@@ -777,6 +778,7 @@ public class XUserREST {
 	@Produces({ "application/xml", "application/json" })
 	@PreAuthorize("@rangerPreAuthSecurityHandler.isAPIAccessible(\"" + RangerAPIList.UPDATE_X_MODULE_DEF_PERMISSION + "\")")
 	public VXModuleDef updateXModuleDefPermission(VXModuleDef vXModuleDef) {
+		xUserMgr.checkAdminAccess();
 		return xUserMgr.updateXModuleDefPermission(vXModuleDef);
 	}
 
@@ -786,6 +788,7 @@ public class XUserREST {
 	public void deleteXModuleDefPermission(@PathParam("id") Long id,
 			@Context HttpServletRequest request) {
 		boolean force = true;
+		xUserMgr.checkAdminAccess();
 		xUserMgr.deleteXModuleDefPermission(id, force);
 	}
 
@@ -827,6 +830,7 @@ public class XUserREST {
 	@PreAuthorize("@rangerPreAuthSecurityHandler.isAPIAccessible(\"" + RangerAPIList.CREATE_X_USER_PERMISSION + "\")")
 	public VXUserPermission createXUserPermission(
 			VXUserPermission vXUserPermission) {
+		xUserMgr.checkAdminAccess();
 		return xUserMgr.createXUserPermission(vXUserPermission);
 	}
 
@@ -844,6 +848,7 @@ public class XUserREST {
 	@PreAuthorize("@rangerPreAuthSecurityHandler.isAPIAccessible(\"" + RangerAPIList.UPDATE_X_USER_PERMISSION + "\")")
 	public VXUserPermission updateXUserPermission(
 			VXUserPermission vXUserPermission) {
+		xUserMgr.checkAdminAccess();
 		return xUserMgr.updateXUserPermission(vXUserPermission);
 	}
 
@@ -853,6 +858,7 @@ public class XUserREST {
 	public void deleteXUserPermission(@PathParam("id") Long id,
 			@Context HttpServletRequest request) {
 		boolean force = true;
+		xUserMgr.checkAdminAccess();
 		xUserMgr.deleteXUserPermission(id, force);
 	}
 
@@ -889,6 +895,7 @@ public class XUserREST {
 	@PreAuthorize("@rangerPreAuthSecurityHandler.isAPIAccessible(\"" + RangerAPIList.CREATE_X_GROUP_PERMISSION + "\")")
 	public VXGroupPermission createXGroupPermission(
 			VXGroupPermission vXGroupPermission) {
+		xUserMgr.checkAdminAccess();
 		return xUserMgr.createXGroupPermission(vXGroupPermission);
 	}
 
@@ -906,6 +913,7 @@ public class XUserREST {
 	@PreAuthorize("@rangerPreAuthSecurityHandler.isAPIAccessible(\"" + RangerAPIList.UPDATE_X_GROUP_PERMISSION + "\")")
 	public VXGroupPermission updateXGroupPermission(
 			VXGroupPermission vXGroupPermission) {
+		xUserMgr.checkAdminAccess();
 		return xUserMgr.updateXGroupPermission(vXGroupPermission);
 	}
 
@@ -915,6 +923,7 @@ public class XUserREST {
 	public void deleteXGroupPermission(@PathParam("id") Long id,
 			@Context HttpServletRequest request) {
 		boolean force = true;
+		xUserMgr.checkAdminAccess();
 		xUserMgr.deleteXGroupPermission(id, force);
 	}
